Rename cytoscape options and make wait cursor
